Add Image attribute to Worker classes, update DAL works regarding Image;minor fixes in money conversion; add ImageHelper.cs to help with conversion between Image and byte[]

+++ CoffeeShopManager/GUI/frmAddEmployee.cs
@@ -60,7 +60,7 @@ namespace GUI
                 dtoEmp.Phone = txtPhone.Text;
                 dtoEmp.Account.Email = txtEmail.Text;
                 dtoEmp.Account.PassWord = txtPassword.Text;
-                dtoEmp.Salary = double.Parse(txtSalary.Text);
+                dtoEmp.Salary = decimal.Parse(txtSalary.Text);
                 if (DateTime.TryParse(txtDayJoin.Text + "/" + txtMonthJoin.Text + "/" + txtYearJoin.Text, out tmp))
                 {
                     dtoEmp.DateOfJoin = tmp;
